# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/typeAb/edit/{id}": {
            "put": {
                "tags": [
                    "TypeAb"
                ],
                "summary": "Update a TypeAb",
                "operationId": "2e5edfd52cb3e1028e11744cf85203a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the TypeAb",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "En cours"
                                    },
                                    "montant": {
                                        "type": "string",
                                        "example": "En cours"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "En cours"
                                    },
                                    "duree": {
                                        "type": "string",
                                        "example": "En cours"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "En cours"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/typeAb/create": {
            "post": {
                "tags": [
                    "TypeAb"
                ],
                "summary": "Create a new TypeAb",
                "operationId": "c2a34a91124f8cba0d72bceec3fea538",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "montant": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "duree": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "TypeAb created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/typeAb/show/{id}": {
            "get": {
                "tags": [
                    "TypeAb"
                ],
                "summary": "Get a specific resource",
                "operationId": "81ff0cf746dbca4d8eddd7fc36dbc6d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": " id de TypeAb",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            }
        },
        "/api/v1/typeAb/delete/{id}": {
            "delete": {
                "tags": [
                    "typeAb"
                ],
                "summary": "Delete a typeAb",
                "operationId": "3955074a58684a96749b21282695ce0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    }
}